combine key and client tests
